Refactor monetário: migrar Double restantes para BigDecimal

+++ app/src/main/java/com/migueldk17/breeze/ui/features/adicionarconta/models/DadosContaUI.kt
@@ -2,6 +2,7 @@ package com.migueldk17.breeze.ui.features.adicionarconta.models
 
 import androidx.compose.ui.graphics.Color
 import com.github.migueldk17.breezeicons.icons.BreezeIconsType
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class DadosContaUI(
@@ -9,12 +10,12 @@ data class DadosContaUI(
     val icone: BreezeIconsType,
     val corIcone: Color,
     val corCard: Color,
-    val valor: Double,
+    val valor: BigDecimal,
     val categoria: String,
     val subCategoria: String,
-    val valorParcela: Double,
+    val valorParcela: BigDecimal,
     val totalParcelas: Int,
     val data: LocalDate,
     val isParcelada: Boolean,
-    val taxaJuros: Double
+    val taxaJuros: BigDecimal
 )
